requirements.txt
PyPDF2
pyyaml
pycryptodome
